unified alfa 1 with 1.0.1-a1

+++ app/Http/Controllers/DocumentController.php
@@ -11,6 +11,8 @@ class DocumentController extends Controller
 {
     public function destroy($id) {
         $documentPhoto = Document::find($id);
+        if( !$documentPhoto->canBeDeleted() )
+            return redirect()->back();
         Helper::deleteImage($documentPhoto->image_path, $thumbImage='');
         $deleted = $documentPhoto->delete();
         return redirect()->back();
